feat: account token
Also, there is a case of a first time user trying to get out of the new
account form page. If the user goes back, the account token is not set
and the user is effectively logged out of the system, as there is no use
of a citizen/notary not having an account.