feat: hashchange
Randomly found this reddit thread https://www.reddit.com/r/solidjs/comments/10did00/creating_spa/ which states that when trying to access SPAs not through the root directory, the site does nothing. Making the router run on hashmode (https://github.com/solidjs/solid-router#hash-mode-router) fixes things. This was not a problem when the app was developed locally (with Vite).